cleaned up interface between billcore and net.
extended encryption interface to allow encrypted connections initiated
from the server.

made it possible to specify an encryption module for the billing server
connection. it now uses "vieenc" by default, so that must be loaded to
connect to any biller. it should detect if the biller supports
encryption or not. (all of this is untested; I expect there to be a few
bugs to work out when it gets used.)