Remove AutoAdd from auth_file. I never liked this feature. It will be replaced by a real password cache in the future.
Add ?local_password as alias for ?passwd, and deprecate ?passwd. I hope
the new name is less confusing.

Add auth_file option: RequireAuthenticationToSetPassword, which makes
?local_password require the user to be authenticated already (either by
auth_file, or a billing server). This defaults to on (in the code),
although the passwd.conf in distributed versions will have it explicitly
set off, since it makes it easier to get started using a local zone.
There's a big comment explaining why you should remove the line in the
conf file and why, and what you should do instead of relying on it.

Add /?set_local_password, which is meant to be used if you're not using
a billing server but you still want to enable
RequireAuthenticationToSetPassword (which is a good idea). It is used by
a sysop on another player, and sets their local password to the password
they used to log in to this session.