this is a revamp of the new freqman code that dr brain wrote. I wrote it in ACE originally (cleaned it up a bit here, but there are still some noticeable artifacts) basically I rewrote the thing almost from scratch (though I'm pretty sure there are parts that are verbatim from brain's code and/or fm_normal before it) the goal was to make the code a little cleaner, fix some fundamental problems with the workflow, fix a memory leak/failure-to-LLRemove that could not be isolated, and add some additional functionality
the interface should be binary-compatible with modules written for the original freqman rewrite.
cb_preshipfreqchange should not be used by custom modules. it was solely to allow freqman to preempt other modules in getting a shipchange callback due to load order snafu

this module ended up replacing the default branch's freqman for what Hockey Zone and Hyperspace uses for over a year now, it seems fairly stable but it seems to have a bug with the balancer code, where if two opposing teams are uneven, favoring the lower frequency, and the balancer settings are normal, it will still put players on the lower frequency if it's not full. however this code should repair the other balancer issue from the other freqman which was a memory leak (#79)

addresses issue 79