ASSS is multithreaded, some of these threads end up calling the Python C API. The Python C API may only be called by a single thread unless you initialize and acquire the GIL properly. See: https://docs.python.org/2/c-api/init.html#non-python-created-threads
This change adds the necessary locking to prevent the many race conditions that plague pymod.

3 macro's are added:

* NEEDS_GIL : local variables needed for the functions that lock
* GI_LOCK : acquire the GIL
* GI_UNLOCK : release the GIL

The code for the lock is a bit complex because python requires 2 different sets of functions for 1) the main thread that spawned python and 2) the other threads. The locking must also be recursive (PyGILState_Ensure already works recursively)